pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lineup-builder"
version = "0.1.0"
description = "Desktop GUI app for building DJ event lineups and generating formatted Discord/plain-text output"
requires-python = ">=3.11"
dependencies = [
    "dearpygui",
    "pyyaml",
    "Pillow",
]

[project.scripts]
lineup-builder = "main:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "watchdog",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
